+++ components/ShowActivity.js
@@ -1,0 +1,14 @@
+import React, { useState, useEffect } from 'react'
+import Header from './Header'
+import RandomActivity from './RandomActivity'
+
+const ShowActivity = () => {
+	return (
+		<>
+			<Header />
+			<RandomActivity />
+		</>
+	)
+}
+
+export default ShowActivity
